backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic Settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Vector Database
qdrant-client==1.7.3

# Embeddings and NLP
sentence-transformers==2.3.1
transformers==4.36.2
torch==2.1.2

# LLM
google-generativeai==0.3.2

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
numpy==1.26.3

# Environment Variables
python-dotenv==1.0.0

# Development
pytest==7.4.4
httpx==0.26.0
